+++ api/src/main/kotlin/org/islandoftex/arara/api/localization/Messages.kt
@@ -19,7 +19,8 @@ internal annotation class AraraMessages
  *
  * TODO: Translate ERROR_INTERPRETER_WRONG_RETURN_TYPE,
  *   ERROR_INTERPRETER_COMMAND_UNSUCCESSFUL_EXIT,
- *   ERROR_INTERPRETER_USER_REQUESTED_HALT
+ *   ERROR_INTERPRETER_USER_REQUESTED_HALT,
+ *   ERROR_EXTRACTOR_INTERPOLATION_FAILURE
  */
 @AraraMessages
 @Suppress("MaxLineLength", "ConstructorParameterNaming", "LongParameterList")
@@ -37,6 +38,7 @@ public open class Messages(
     public val ERROR_EVALUATE_COMPILATION_FAILED: String = "For whatever reason, I could not compile the expression in the provided conditional. This part is quite tricky, since it involves aspects of the underlying expression language. I will do my best to help you in any way I can.",
     public val ERROR_EVALUATE_NOT_BOOLEAN_VALUE: String = "The conditional evaluation was expecting a boolean value as result. This should be an easy fix. Just make sure the conditional evaluation resolves to a boolean value in the end.",
     public val ERROR_EXTRACTOR_IO_ERROR: String = "There was an IO error while I was trying to extract the directives. I have no idea why it failed, though. Perhaps the file was moved or deleted before or during the extraction operation. Or maybe I do not have the proper permissions to read the file.",
+    public val ERROR_EXTRACTOR_INTERPOLATION_FAILURE: String = "Your 'options' field contains an entry the MVEL parser is unable to cope with. Please revise your options not to contain any malicious '@{…}' parts.",
     public val ERROR_FILETYPE_NOT_A_FILE: String = "The \"filetype\" method requires a file, not a directory. It looks like \"%s\" does not appear to be a file at all. If you need to perform tasks on a directory, you could use a couple of methods from the Java API.",
     public val ERROR_FILETYPE_UNKNOWN_EXTENSION: String = "I cannot recognize \"%s\" as a default extension. If you want to define a new file type, make sure to provide the extension and pattern. These are the default extensions: %s",
     public val ERROR_GETAPPLICATIONPATH_ENCODING_EXCEPTION: String = "There was an encoding problem while trying to obtain the application path. There is nothing much I can do about it.",