+++ CHANGELOG.md
@@ -17,6 +17,10 @@ from version 5.0.0 on.
   by calling `getSession().get('arg:key')` if you called arara with
   `arara -P key=value myfile.tex`.
   (see #38)
+* `arara` now expands `@{}` orb tags within `options` parameter to directives.
+  Supported methods are `getBasename`, `getSession` and `getOriginalReference`
+  with their documented meanings.
+  (see #38, #46)
 * A new safe mode (`-s` or `--safe-run`) protects the system by disallowing
   certain user actions.
   (see #18)
@@ -26,6 +30,7 @@ from version 5.0.0 on.
     (although rules are still allowed to construct arbitrary commands using
     a `return getCommand(…)` approach so this only disallows arbitrary system
     commands that would not get logged and are thus invisible to the user)
+  * the `options` parameter does not expand orb tags in any directive.
 * `-dev` formats are now available within the `latexmk` rule.
   (see #48)